Исправление замечаний. - перенес функции для загрузки данных в один интерфейс PictureOfTheDayAPI и удалил лишние имплементации ретрофита. Теперь viewModel каждого экрана использует один PictureOfTheDayRetrofitImpl для получения своих данных. - определил дата классы serverResponseData в отдельные файлы.
