+++ app/src/main/java/com/geekbrains/pictureoftheday/model/mars/MarsData.kt
@@ -1,5 +1,7 @@
 package com.geekbrains.pictureoftheday.model.mars
 
+import com.geekbrains.pictureoftheday.model.mars.serverResponseData.MarsServerResponseData
+
 sealed class MarsData {
     data class Success(val serverResponseData: MarsServerResponseData) : MarsData()
     data class Error(val error: Throwable) : MarsData()
