--- app/src/main/java/com/geekbrains/pictureoftheday/model/mars/MarsServerResponseData.kt
@@ -1,31 +1,0 @@
-package com.geekbrains.pictureoftheday.model.mars
-
-import com.google.gson.annotations.SerializedName
-
-class MarsServerResponseData(
-    @SerializedName("photos") val photos: List<Photo>
-) {
-    data class Photo(
-        @SerializedName("camera") val camera: Camera,
-        @SerializedName("earth_date") val earthDate: String,
-        @SerializedName("id") val id: Int,
-        @SerializedName("img_src") val imgSrc: String,
-        @SerializedName("rover") val rover: Rover,
-        @SerializedName("sol") val sol: Int
-    ) {
-        data class Camera(
-            @SerializedName("full_name") val fullName: String,
-            @SerializedName("id") val id: Int,
-            @SerializedName("name") val name: String,
-            @SerializedName("rover_id") val roverId: Int
-        )
-
-        data class Rover(
-            @SerializedName("id") val id: Int,
-            @SerializedName("landing_date") val landingDate: String,
-            @SerializedName("launch_date") val launchDate: String,
-            @SerializedName("name") val name: String,
-            @SerializedName("status") val status: String
-        )
-    }
-}
