+++ app/src/main/java/com/geekbrains/pictureoftheday/model/mars/serverResponseData/MarsServerResponseData.kt
@@ -1,0 +1,7 @@
+package com.geekbrains.pictureoftheday.model.mars.serverResponseData
+
+import com.google.gson.annotations.SerializedName
+
+class MarsServerResponseData(
+    @SerializedName("photos") val photos: List<Photo>
+)
