--- app/src/main/java/com/geekbrains/pictureoftheday/model/meteorite/MeteoriteAPI.kt
@@ -1,15 +1,0 @@
-package com.geekbrains.pictureoftheday.model.meteorite
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface MeteoriteAPI {
-
-    @GET("neo/rest/v1/feed")
-    fun getMeteorite(
-        @Query("START_DATE") startDate: String,
-        @Query("end_date") endDate: String,
-        @Query("api_key") apiKey: String
-    ): Call<MeteoriteServerResponseData>
-    }
