+++ app/src/main/java/com/geekbrains/pictureoftheday/model/meteorite/MeteoriteData.kt
@@ -1,5 +1,7 @@
 package com.geekbrains.pictureoftheday.model.meteorite
 
+import com.geekbrains.pictureoftheday.model.meteorite.serverResponseData.MeteoriteServerResponseData
+
 sealed class MeteoriteData {
     data class Success(val serverResponseData: MeteoriteServerResponseData) : MeteoriteData()
     data class Error(val error: Throwable) : MeteoriteData()
