+++ app/src/main/java/com/geekbrains/pictureoftheday/model/pod/PODData.kt
@@ -1,0 +1,7 @@
+package com.geekbrains.pictureoftheday.model.pod
+
+sealed class PODData {
+    data class Success(val serverResponseData: PODServerResponseData) : PODData()
+    data class Error(val error: Throwable) : PODData()
+    data class Loading(val progress: Int?) : PODData()
+}
