--- app/src/main/java/com/geekbrains/pictureoftheday/model/pod/PODRetrofitImp.kt
@@ -1,37 +1,0 @@
-package com.geekbrains.pictureoftheday.model.pod
-
-import com.google.gson.GsonBuilder
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.Response
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class PODRetrofitImp {
-    private val baseUrl = "https://api.nasa.gov/"
-
-    fun getRetrofitImp(): PictureOfTheDayAPI {
-        val podRetrofit = Retrofit.Builder()
-            .baseUrl(baseUrl)
-            .addConverterFactory(GsonConverterFactory.create(GsonBuilder().setLenient().create()))
-            .client(createOkHttpClient(PODInterceptor()))
-            .build()
-
-        return podRetrofit.create(PictureOfTheDayAPI::class.java)
-    }
-
-    private fun createOkHttpClient(interceptor: Interceptor): OkHttpClient {
-        val httpClient = OkHttpClient.Builder()
-        httpClient.addInterceptor(interceptor)
-        httpClient.addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-
-        return httpClient.build()
-    }
-
-    inner class PODInterceptor : Interceptor {
-        override fun intercept(chain: Interceptor.Chain): Response {
-            return chain.proceed(chain.request())
-        }
-    }
-}
