--- app/src/main/java/com/geekbrains/pictureoftheday/model/pod/PictureOfTheDayAPI.kt
@@ -1,10 +1,0 @@
-package com.geekbrains.pictureoftheday.model.pod
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface PictureOfTheDayAPI {
-    @GET ("planetary/apod")
-    fun getPictureOfTheDay(@Query("api_key") apiKey: String): Call<PODServerResponseData>
-}
