--- app/src/main/java/com/geekbrains/pictureoftheday/model/pod/PictureOfTheDayData.kt
@@ -1,7 +1,0 @@
-package com.geekbrains.pictureoftheday.model.pod
-
-sealed class PictureOfTheDayData {
-    data class Success(val serverResponseData: PODServerResponseData) : PictureOfTheDayData()
-    data class Error(val error: Throwable) : PictureOfTheDayData()
-    data class Loading(val progress: Int?) : PictureOfTheDayData()
-}
