+++ app/src/main/java/com/geekbrains/pictureoftheday/model/spaceweather/SpaceWeatherData.kt
@@ -1,5 +1,7 @@
 package com.geekbrains.pictureoftheday.model.spaceweather
 
+import com.geekbrains.pictureoftheday.model.spaceweather.serverResponseData.SpaceWeatherServerResponseData
+
 sealed class SpaceWeatherData {
     data class Success(val serverResponseData: List<SpaceWeatherServerResponseData>) : SpaceWeatherData()
     data class Error(val error: Throwable) : SpaceWeatherData()
