+++ app/src/main/java/com/geekbrains/pictureoftheday/model/spaceweather/serverResponseData/KpIndex.kt
@@ -1,0 +1,7 @@
+package com.geekbrains.pictureoftheday.model.spaceweather.serverResponseData
+
+import com.google.gson.annotations.SerializedName
+
+data class KpIndex(
+    @SerializedName("kpIndex") val kpIndex: Int,
+)
