+++ app/src/main/java/com/geekbrains/pictureoftheday/model/spaceweather/serverResponseData/SpaceWeatherServerResponseData.kt
@@ -1,11 +1,7 @@
-package com.geekbrains.pictureoftheday.model.spaceweather
+package com.geekbrains.pictureoftheday.model.spaceweather.serverResponseData
 
 import com.google.gson.annotations.SerializedName
 
 class SpaceWeatherServerResponseData(
     @SerializedName("allKpIndex") val allKpIndex: List<KpIndex>
-) {
-    data class KpIndex(
-        @SerializedName("kpIndex") val kpIndex: Int,
-    )
-}
+)
