+++ app/src/main/java/com/geekbrains/pictureoftheday/viewModel/meteorite/MeteoriteViewModel.kt
@@ -4,21 +4,18 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.geekbrains.pictureoftheday.BuildConfig
+import com.geekbrains.pictureoftheday.model.PictureOfTheDayRetrofitImp
 import com.geekbrains.pictureoftheday.model.meteorite.MeteoriteData
-import com.geekbrains.pictureoftheday.model.meteorite.MeteoriteRetrofitImp
-import com.geekbrains.pictureoftheday.model.meteorite.MeteoriteServerResponseData
+import com.geekbrains.pictureoftheday.model.meteorite.serverResponseData.MeteoriteServerResponseData
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-import java.util.*
 
-const val date = "2022-03-07"
+const val date = "2022-03-10"
 
 class MeteoriteViewModel(
     private val liveDataForViewToObserve: MutableLiveData<MeteoriteData> = MutableLiveData(),
-    private val retrofitImp: MeteoriteRetrofitImp = MeteoriteRetrofitImp()
+    private val retrofitImp: PictureOfTheDayRetrofitImp = PictureOfTheDayRetrofitImp()
 ) : ViewModel() {
 
     fun getData(): LiveData<MeteoriteData> {
@@ -33,7 +30,7 @@ class MeteoriteViewModel(
         if (apiKey.isBlank()) {
             MeteoriteData.Error(Throwable("You need API key"))
         } else {
-            retrofitImp.getRetrofitImp().getMeteorite(date,date , apiKey).enqueue(object :
+            retrofitImp.getRetrofitImp().getMeteorite(date, date, apiKey).enqueue(object :
                 Callback<MeteoriteServerResponseData> {
                 override fun onResponse(
                     call: Call<MeteoriteServerResponseData>,
