+++ app/src/main/java/com/geekbrains/pictureoftheday/viewModel/pod/PODViewModel.kt
@@ -4,31 +4,31 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.geekbrains.pictureoftheday.BuildConfig
-import com.geekbrains.pictureoftheday.model.pod.PODRetrofitImp
+import com.geekbrains.pictureoftheday.model.PictureOfTheDayRetrofitImp
+import com.geekbrains.pictureoftheday.model.pod.PODData
 import com.geekbrains.pictureoftheday.model.pod.PODServerResponseData
-import com.geekbrains.pictureoftheday.model.pod.PictureOfTheDayData
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class PictureOfTheDayViewModel(
-    private val liveDataForViewToObserve: MutableLiveData<PictureOfTheDayData> = MutableLiveData(),
-    private val retrofitImp: PODRetrofitImp = PODRetrofitImp()
+class PODViewModel(
+    private val liveDataForViewToObserve: MutableLiveData<PODData> = MutableLiveData(),
+    private val retrofitImp: PictureOfTheDayRetrofitImp = PictureOfTheDayRetrofitImp()
 ) : ViewModel() {
 
-    fun getData(): LiveData<PictureOfTheDayData> {
+    fun getData(): LiveData<PODData> {
         sendServerRequest()
         return liveDataForViewToObserve
     }
 
     private fun sendServerRequest() {
-        liveDataForViewToObserve.value = PictureOfTheDayData.Loading(null)
+        liveDataForViewToObserve.value = PODData.Loading(null)
         val apiKey: String = BuildConfig.NASA_API_KEY
 
         if (apiKey.isBlank()) {
-            PictureOfTheDayData.Error(Throwable("You need API key"))
+            PODData.Error(Throwable("You need API key"))
         } else {
-            retrofitImp.getRetrofitImp().getPictureOfTheDay(apiKey).enqueue(object :
+            retrofitImp.getRetrofitImp().getPOD(apiKey).enqueue(object :
                 Callback<PODServerResponseData> {
                 override fun onResponse(
                     call: Call<PODServerResponseData>,
@@ -36,21 +36,21 @@ class PictureOfTheDayViewModel(
                 ) {
                     if (response.isSuccessful && response.body() != null) {
                         liveDataForViewToObserve.value =
-                            PictureOfTheDayData.Success(response.body()!!)
+                            PODData.Success(response.body()!!)
                     } else {
                         val message = response.message()
                         if (message.isNullOrEmpty()) {
                             liveDataForViewToObserve.value =
-                                PictureOfTheDayData.Error(Throwable("Unidentified error"))
+                                PODData.Error(Throwable("Unidentified error"))
                         } else {
                             liveDataForViewToObserve.value =
-                                PictureOfTheDayData.Error(Throwable(message))
+                                PODData.Error(Throwable(message))
                         }
                     }
                 }
 
                 override fun onFailure(call: Call<PODServerResponseData>, t: Throwable) {
-                    liveDataForViewToObserve.value = PictureOfTheDayData.Error(t)
+                    liveDataForViewToObserve.value = PODData.Error(t)
                 }
             })
         }
